requirements.txt
flask>=3.0.3
requests>=2.32.3
pandas>=2.2.2
psycopg2-binary>=2.9.9
openpyxl>=3.1.5
reportlab>=4.2.2
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-flask>=1.3.0
ruff>=0.3.0
